+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* add type-level digit string parsing at `Data.Type.Symbol.Natural`
+
 ## 0.1.0 (2024-04-13)
 Initial release.